Allow a single decimal place for font size setting
On windows hidpi displays this IS required to get nice sizing, eg
when the display scaling is set to a non-integer scaling factor
(ie 150%)

+++ src/app/options/qgsoptions.h
@@ -113,7 +113,7 @@ class APP_EXPORT QgsOptions : public QgsOptionsDialogBase, private Ui::QgsOption
     void selectProjectOnLaunch();
 
     //! Slot to select the default font point size for app
-    void spinFontSize_valueChanged( int fontSize );
+    void spinFontSize_valueChanged(double fontSize );
 
     //! Slot to set font family for app to Qt default
     void mFontFamilyRadioQt_released();
